make clear, what "/" in index.txt lines will do
We might allow /-prefixed path in the future with some special
meaning, so better to disallow right now.

Paths with slashes mean, that in the context of this spec they
are part of an icon set. We need to disallow clients to load
some wrong icon just because the basename matches.